+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Change
 - Remove `getTalliedMarks` function
 - Add `createCopy` argument to `getMarkData`
-- Add `getAllMarkData` function
 - Rename `start` to `begin`
+- Rename `getMarkData` to `getTimingData`
+- Add `getAllTimingData` function
+- Rename `getPendingMarks` to `getPending`
 
 ## [0.2.2] - 2018-08-15
 ### Fixed